refactor(core): remove legacy color tokens
BREAKING CHANGE remove legacy color tokens
