+++ packages/core/js/index.js
@@ -1,6 +1,5 @@
 export { default as accessibility } from './accessibility.js'
-export { default as colors } from './colors.js'
 export { default as layout } from './layout.js'
 export { default as type } from './type.js'
 export { default as motion } from './motion.js'
-export * from './colors.2020.js'
+export * from './colors.js'
